+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+- Improve formatter to not rely on regex
+- Support extglob
 - Support file descriptors 3+.
 - Full support for coprocs.
 - `wait` command now properly waits.